Nummer sicher

+++ aufgabenblatt_2/aufgabe4/fourier_opencl.cpp
@@ -56,7 +56,7 @@ std::pair<std::vector<float>, std::vector<float>> perform_fft_gpu(cl::Context& c
     for (int i = 0; i < block_size / 2; ++i) {
         frequencies[i] = i * static_cast<float>(sample_rate) / block_size;
     }
-    std::vector<float> fft_blocks(num_blocks * (block_size / 2));
+    std::vector<float> fft_blocks((size_t)num_blocks * (block_size / 2));
 
     cl::Buffer data_buffer(context, CL_MEM_READ_ONLY | CL_MEM_COPY_HOST_PTR, sizeof(float) * data.size(), const_cast<float*>(data.data()));
     cl::Buffer fft_buffer(context, CL_MEM_WRITE_ONLY, sizeof(float) * fft_blocks.size());
